ComfyUI/custom_nodes/requirements.txt
#torch
#torchsde
#torchvision
einops
transformers>=4.25.1
safetensors>=0.3.0
aiohttp
accelerate
pyyaml
Pillow
scipy
tqdm
psutil
llama-cpp-python
tqdm
nltk
matplotlib
#aphrodite